requirements.txt
# Core dependencies 
pandas>=2.0.0 
tabulate==0.9.0 
 
# Database (PostgreSQL for cloud) 
psycopg2-binary>=2.9.0 
 
# Web framework (for browser interface) 
flask>=3.0.0 
gunicorn>=21.0.0 
 
# Excel export 
openpyxl>=3.1.0
